design check trait
