+++ src/faebryk/libs/app/checks.py
@@ -7,13 +7,11 @@ import faebryk.library._F as F
 from faebryk.core.graph import Graph, GraphFunctions
 from faebryk.core.module import Module
 from faebryk.core.node import Node
+from faebryk.library.implements_design_check import CheckException
 from faebryk.libs.app.erc import simple_erc
 from faebryk.libs.exceptions import accumulate
 
 logger = logging.getLogger(__name__)
-
-
-class CheckException(Exception): ...
 
 
 class RequiresExternalUsageNotFulfilled(CheckException):
@@ -21,15 +19,6 @@ class RequiresExternalUsageNotFulfilled(CheckException):
         self.nodes = nodes
         super().__init__(
             f"Nodes requiring external usage but not used externally: "
-            f"{', '.join(mif.get_full_name() for mif in nodes)}"
-        )
-
-
-class RequiresPullNotFulfilled(CheckException):
-    def __init__(self, nodes: list[Node]):
-        self.nodes = nodes
-        super().__init__(
-            f"Signals requiring pulls but not pulled: "
             f"{', '.join(mif.get_full_name() for mif in nodes)}"
         )
 
@@ -43,13 +32,13 @@ def check_requires_external_usage(G: Graph):
         raise RequiresExternalUsageNotFulfilled(unfulfilled)
 
 
-def check_requires_pulls(G: Graph):
-    unfulfilled = []
-    for node, trait in GraphFunctions(G).nodes_with_trait(F.requires_pulls):
-        if not trait.fulfilled:
-            unfulfilled.append(node)
-    if unfulfilled:
-        raise RequiresPullNotFulfilled(unfulfilled)
+def check_design(G: Graph):
+    with accumulate(CheckException) as accumulator:
+        with accumulator.collect():
+            for _, trait in GraphFunctions(G).nodes_with_trait(
+                F.implements_design_check
+            ):
+                trait.check()
 
 
 def run_checks(app: Module, G: Graph):
@@ -58,6 +47,6 @@ def run_checks(app: Module, G: Graph):
     with accumulate(CheckException) as accumulator:
         with accumulator.collect():
             check_requires_external_usage(G)
-            check_requires_pulls(G)
+            check_design(G)
 
     simple_erc(G)
